backend/requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-multipart
python-dotenv